acyclic: added string term in movebyte before handler is called
This allows the handle to work with the commandline as normal C-string.
And fixes an issue where the following happened:

$ reg write 1234
1234: ok

$ reg read 4321
43214: denied

This was possible because "reg read 4321" wasn't null terminated and
still had the "4" from "reg write 1234" in it. The handler used a string
to int function that scanned until a non-number or a null terminator was
found, so the non-visible "4" that was still in the buffer was detected
as valid number in the string.